% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_simulate.R
\name{annotate.multistate.groups}
\alias{annotate.multistate.groups}
\title{Annotate the colData with the multistate group}
\usage{
annotate.multistate.groups(sim)
}
\arguments{
\item{sim}{Simulated QTLExperiment object to annotate}
}
\description{
Annotate the colData with the multistate group
}
