% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_simulate.R
\name{.multistate_pattern_check}
\alias{.multistate_pattern_check}
\title{Check if simulated multistate pattern passes tests}
\usage{
.multistate_pattern_check(pattern, existing)
}
\arguments{
\item{pattern}{multi-state QTL pattern to test}

\item{existing}{list of existing multi-state QTL patterns}
}
\description{
Check if simulated multistate pattern passes tests
}
