% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_functions.R
\name{.resolve_annotations}
\alias{.resolve_annotations}
\title{Get ComplexHeatmap annotations for rows and columns}
\usage{
.resolve_annotations(anns, FUN = rowAnnotation)
}
\arguments{
\item{anns}{data.frame with annotations as columns}

\item{FUN}{function to use to generate ComplexHeatmap annotation object.
HeatmapAnnotation for columns and rowAnnotation for rows.}
}
\description{
Get ComplexHeatmap annotations for rows and columns
}
