% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_functions.R
\name{.resolve_complexheatmap_colors}
\alias{.resolve_complexheatmap_colors}
\title{Color palettes for ComplexHeatmaps and their annotations. If more than 8
continuous or 4 qualitative palettes are requested they will start repeating.}
\usage{
.resolve_complexheatmap_colors(color_by, i = 1)
}
\arguments{
\item{color_by}{array with values to color by}

\item{i}{index of color scheme to select}
}
\description{
Color palettes for ComplexHeatmaps and their annotations. If more than 8
continuous or 4 qualitative palettes are requested they will start repeating.
}
