% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_functions.R
\name{.resolve_plot_colors}
\alias{.resolve_plot_colors}
\title{Color palettes for ggplot and other (non-ComplexHeatmap) plotting functions.}
\usage{
.resolve_plot_colors(plot_out, color_by, color_by_name, fill = FALSE)
}
\arguments{
\item{plot_out}{plot object}

\item{color_by}{array with values to color by}

\item{color_by_name}{name of color_by array}

\item{fill}{Logical indicating if the palette is for fill or color (e.g.
scale_fill_brewer vs scale_color_brewer).}
}
\description{
Color palettes for ggplot and other (non-ComplexHeatmap) plotting functions.
}
