% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getComplete}
\alias{getComplete}
\alias{getTopHits}
\title{Filter QTLExperiment based on missing data}
\usage{
getComplete(qtle, n = 1, verbose = FALSE)

getTopHits(
  qtle,
  mode = c("global", "state"),
  assay = "pvalues",
  assay_sig = "significant",
  verbose = FALSE
)
}
\arguments{
\item{qtle}{A `QTLExperiment` object}

\item{n}{Number (or percent if n < 1) of states requiring non-null values}

\item{verbose}{logical. Whether to print progress messages.}

\item{mode}{global/state to specify if the top hit per feature is desired 
from across all states or for each state.}

\item{assay}{The assay containing the test statistic to minimize.}

\item{assay_sig}{The assay containing TRUE/FALSE significance calls for each
QTL test.}
}
\value{
a subset of the `QTLExperiment` object, with only tests with fewer
NAs than specified by n.

a subset of the `QTLExperiment` object, with only tests that are the
top hits for each feature (`mode=global`) or for each feature for each 
state (`mode=state`).
}
\description{
Method to filter \linkS4class{QTLExperiment} objects to remove tests with 
greater than the permitted rate of missing values.

Method to return a subset of a \linkS4class{QTLExperiment} object containing
only the tests that are top hits. Top hits are defined as the test for each 
feature with the  an array of the top QTL for each feature across all states
}
