% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getSignificant}
\alias{getSignificant}
\title{Return an array of the QTL significant in at least one state}
\usage{
getSignificant(qtle, n = 1, assay = "significant", verbose = FALSE)
}
\arguments{
\item{qtle}{`QTLExperiment` object}

\item{n}{Number (or percent if n < 1) of states with significant association}

\item{assay}{The assay containing TRUE/FALSE significance calls for each
QTL test.}

\item{verbose}{logical. Whether to print progress messages.}
}
\description{
Return an array of the QTL significant in at least one state
}
