% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_simulate.R
\name{make.ms.simulation.key}
\alias{make.ms.simulation.key}
\title{Add multi-state clusters to simulation key}
\usage{
make.ms.simulation.key(key, states, k, verbose)
}
\arguments{
\item{key}{matrix with simulation information stored.}

\item{states}{array of state names}

\item{k}{number of multi-state clusters}

\item{verbose}{Logical}
}
\description{
Add multi-state clusters to simulation key
}
