% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_simulate.R
\name{make.simulation.key}
\alias{make.simulation.key}
\title{Generate simulation key}
\usage{
make.simulation.key(params, features, states, types, k, verbose)
}
\arguments{
\item{params}{list of parameters required to simulate betas and beta errors. 
Generated by `qtleEstimate()` or `qtleParams()`.}

\item{features}{array of .feature_ids}

\item{states}{array of state names}

\item{types}{array of test types to include in simulation}

\item{k}{number of multi-state clusters}

\item{verbose}{Logical.}
}
\description{
Generate simulation key
}
