% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSimulationParams}
\alias{plotSimulationParams}
\title{Distribution of estimated simulation parameters}
\usage{
plotSimulationParams(params, n = 1e+05)
}
\arguments{
\item{params}{Estimated simulation parameters from qtleEstimate.}

\item{n}{Number of random values to sample for plots.}
}
\value{
A ggplot2 object
}
\description{
Distribution of estimated simulation parameters
}
\examples{
qtle <- mockQTLE() 
params <- qtleEstimate(qtle, thresh_sig = 0.05, thresh_null = 0.5)
plotSimulationParams(params=params)


}
