% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPairwiseSharing}
\alias{plotPairwiseSharing}
\alias{plotUpSet}
\alias{plotQTLClusters}
\alias{plotCompareStates}
\alias{plotSimulationParams}
\title{Heatmap of pairwise QTL sharing with state-level annotations}
\usage{
plotPairwiseSharing(
  object,
  slot = "pairwiseSharing",
  annotate_rows = NULL,
  annotate_cols = NULL,
  cell_annotate = FALSE,
  col_range = NULL,
  name = "colnames",
  dist.method = "pearson",
  size = 8,
  ...
)

plotUpSet(
  object,
  assay = "significant",
  name = "colnames",
  min_shared = 10,
  min_degree = 2,
  max_degree = NULL,
  annotate_by = NULL,
  comb_order = "comb_size",
  set_order = order(ss)
)

plotQTLClusters(
  object,
  fill_by = "betas",
  FUN = identity,
  min_sig = 1,
  annotate_states = NULL,
  annotate_tests = NULL,
  show_row_names = FALSE,
  state_id = "state_id",
  column_order = NULL,
  row_order = NULL,
  row_km = 0,
  col_km = 0
)

plotCompareStates(
  object,
  x,
  y,
  assay = "betas",
  FUN = identity,
  significance_assay = "significant",
  alpha = 0.2,
  col_both = "#4477AA",
  col_diverging = "#EE6677",
  col_neither = "gray50",
  col_x = "#CCBB44",
  col_y = "#AA3377"
)

plotSimulationParams(params, n = 1e+05)
}
\arguments{
\item{object}{an \code{QTLExperiment} object.}

\item{slot}{Name of slot in metadata list with Pairwise Sharing matrix.}

\item{annotate_rows}{character or array of characters specifying the
column(s) in colData to be plotted as row annotations.}

\item{annotate_cols}{character or array of characters specifying the
column(s) in colData to be plotted as column annotations.}

\item{cell_annotate}{Logical to annoate cells with their values.}

\item{col_range}{Optional range for the color legend}

\item{name}{character specifying the column in colData to use to label rows
and columns. Default is colnames(qtle).}

\item{dist.method}{Distance method used for hierarchical clustering. Valid
values are the supported methods in dist() function.}

\item{size}{numeric scalar giving default font size for plotting theme.}

\item{...}{further arguments passed to \code{\link[Rtsne]{Rtsne}}}

\item{assay}{name of assay to plot.}

\item{min_shared}{minimum number of shared QTL for set to be included.}

\item{min_degree}{minimum degree of sharing for set to be included.}

\item{max_degree}{maximum degree of sharing for set to be included.}

\item{annotate_by}{character or array of characters specifying the
column(s) in colData to be plotted as annotations.}

\item{comb_order}{characters specifying how sets should be ordered. Options
include the set size (set_size), combination size (comb_size), degree (deg).}

\item{set_order}{Array specifying order of states.}

\item{fill_by}{name of assay to use for main heatmap object.}

\item{FUN}{Function to be applied to fill_by assay before plotting (e.g.
identity, abs, log10).}

\item{min_sig}{minimum number of significant states for QTL to be included.}

\item{annotate_states}{character or array of characters specifying the
column(s) in colData to be plotted to annotate states.}

\item{annotate_tests}{character or array of characters specifying the
column(s) in rowData to be plotted to annotate QTL tests.}

\item{show_row_names}{Logical to plot row (i.e. test) names.}

\item{state_id}{colData column to use to label states.}

\item{column_order}{Null for clustering or array to overwrite state order}

\item{row_order}{Null for clustering or array to overwrite test order}

\item{row_km}{Set k for k-means clustering of tests.}

\item{col_km}{Set k for k-means clustering of states}

\item{x}{Name of state for x-axis}

\item{y}{Name of state for y-axis}

\item{significance_assay}{name of assay with TRUE/FALSE significance calls.}

\item{alpha}{Transparency.}

\item{col_both}{Color for tests significant in both states.}

\item{col_diverging}{Color for tests significant in both states, with
diverging effect sizes.}

\item{col_neither}{Color for null tests.}

\item{col_x}{Color for tests significant in the x-axis state only.}

\item{col_y}{Color for tests significant in the y-axis state only.}

\item{params}{Estimated simulation parameters from qtleEstimate.}

\item{n}{Number of random values to sample for plots.}
}
\value{
Returns a \code{ComplexHeatmap} object.

Returns a \code{ComplexHeatmap} object.

Returns a \code{ComplexHeatmap} object.

Returns a list containing the counts for each color category
        and the plot object.

A ggplot2 object
}
\description{
Methods to plot at heatmap of the pairwise sharing of QTL as calculated
by `runPairwiseSharing`.

Convenient method to plot an UpSet plot showing the number of QTL that are
significant in sets of states.

Convenience method for plotting values from any assay specified by fill_by
across states and tests.

Convenience method for comparing the assay value, specified by assay,
between two states.
}
