% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimate.R, R/sim_performance.R,
%   R/sim_simulate.R
\name{qtleEstimate}
\alias{qtleEstimate}
\alias{qtleEstimate.QTLExperiment}
\alias{qtleEstimate.list}
\alias{simPerformance}
\alias{qtleSimulate}
\title{Estimate parameters from real data for simulating multi-state QTL summary statistics}
\usage{
qtleEstimate(
  data,
  assay = "pvalues",
  thresh_sig = 0.001,
  thresh_null = 0.1,
  verbose = TRUE
)

\method{qtleEstimate}{QTLExperiment}(
  data,
  assay = "pvalues",
  thresh_sig = 0.001,
  thresh_null = 0.1,
  verbose = TRUE
)

\method{qtleEstimate}{list}(
  data,
  assay = "pvalues",
  thresh_sig = 0.01,
  thresh_null = 0.1,
  verbose = TRUE
)

simPerformance(qtle, assay = "significant")

qtleSimulate(
  params = qtleParams(),
  ntests = 100,
  nfeatures = NULL,
  nstates = 5,
  global = 0.5,
  multi = 0,
  unique = 0,
  k = 2,
  beta_sd = 0.1,
  lfsr = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A `QTLExperiment` object or named list containing "betas" and
"errors" matrices.}

\item{assay}{Name of the `multistateQTL` assay containing the significance calls}

\item{thresh_sig}{Max threshold (pval/lfsr) for calling tests as significant.}

\item{thresh_null}{Min threshold (pval/lfsr) for calling tests as null.}

\item{verbose}{Logical.}

\item{qtle}{A `multistateQTL` object.}

\item{params}{list of parameters required to simulate betas and beta errors. 
Generated by `qtleEstimate()` or `qtleParams()`.}

\item{ntests}{number of QTL tests}

\item{nfeatures}{number of QTL features to simulate tests for, NULL mean 
nfeatures = ntests.}

\item{nstates}{number of states}

\item{global}{percent of QTL tests with significant effects shared across all states}

\item{multi}{percent of QTL tests with significant effects shared across a subset of states.}

\item{unique}{percent of QTL tests with significant effects in only one state}

\item{k}{number of multi-state clusters or an array with the cluster assignments.}

\item{beta_sd}{The desired standard deviation or an array of sd equal to the
length of states for sampling beta values for each state.}

\item{lfsr}{Logical to calculate lfsr using mashr_1by1.}
}
\description{
Estimate parameters from real data for simulating multi-state QTL summary statistics

Performance metrics for multistateQTL simulations

Create simulated multistateQTL data for testing purposes
}
\details{
The simulation consists of user defined number of equal numbers of four different
types of effects: null, equal among conditions, present only in
first condition, independent across conditions

The simulation consists of user defined number of equal numbers of four different
types of effects: null, equal among conditions, present only in
first condition, independent across conditions
}
\examples{
default_params <- qtleParams()
qtleEstimate(mockQTLE())

sim <- qtleSimulate()
sim <- callSignificance(sim, assay="lfsrs", thresh=0.1)
simPerformance(sim)


qtleSimulate(ntests=100, nstates=5, global=0.1, multi=0.2, unique=0.05)

}
