% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSignificantFeatures.R
\name{runSignificantFeatures}
\alias{runSignificantFeatures}
\title{Function to summarize QTL summary statistics by feature ID}
\usage{
runSignificantFeatures(object, assay = "significant")
}
\arguments{
\item{object}{A \linkS4class{QTLExperiment} object with multiple
QTL tests (i.e., rows) for at least one feature.}

\item{assay}{Assay containing T/F significance calls for each test.}
}
\description{
This function pulls the top QTL for each feature (e.g., gene) and
outputs a **new** \linkS4class{QTLExperiment}. Note that the
output from this function should be assigned to a new variable as it will
overwrite if assigned to the same variable.
}
