% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_performance.R
\name{simPerformance}
\alias{simPerformance}
\title{Performance metrics for multistateQTL simulations}
\usage{
simPerformance(qtle, assay = "significant")
}
\arguments{
\item{qtle}{A `multistateQTL` object.}

\item{assay}{Name of the `multistateQTL` assay containing the significance calls}
}
\value{
description
}
\description{
Performance metrics for multistateQTL simulations
}
\examples{
sim <- qtleSimulate()
sim <- callSignificance(sim, assay="lfsrs", thresh=0.1)
simPerformance(sim)

}
