% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_simulate.R
\name{simulateBetas}
\alias{simulateBetas}
\title{Simulate beta values}
\usage{
simulateBetas(key, params, states, beta_sd, verbose)
}
\arguments{
\item{key}{data.frame containing information about how to simulate multistateQTL. 
Generated by `make.simulation.key()`.}

\item{params}{list of parameters required to simulate betas and beta errors. 
Generated by `qtleEstimate()` or `qtleParams()`.}

\item{states}{array of state names}

\item{beta_sd}{The desired standard deviation or an array of sd equal to the
length of states for sampling beta values for each state.}

\item{verbose}{Logical.}
}
\description{
Simulate beta values
}
