% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_simulate.R
\name{simulateErrors}
\alias{simulateErrors}
\title{Simulate error values}
\usage{
simulateErrors(key, params, sim_betas, verbose)
}
\arguments{
\item{key}{data.frame containing information about how to simulate multistateQTL.
Generated by `make.simulation.key()`.}

\item{params}{list of parameters required to simulate betas and beta errors.
Generated by `qtleEstimate()` or `qtleParams()`.}

\item{sim_betas}{Output from `simulateBetas()`.}

\item{verbose}{Logical.}
}
\description{
Simulate error values
}
