% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistateQTL-package.R
\docType{package}
\name{multistateQTL-package}
\alias{multistateQTL}
\alias{multistateQTL-package}
\title{multistateQTL: Toolkit for the analysis of multi-state QTL data}
\description{
A collection of tools for doing various analyses of multi-state QTL data, with a focus on visualization and interpretation. The package 'multistateQTL' contains functions which can remove or impute missing data, identify significant associations, as well as categorise features into global, multi-state or unique. The analysis results are stored in a 'QTLExperiment' object, which is based on the 'SummarisedExperiment' framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dunstone-a/multistateQTL}
  \item Report bugs at \url{https://github.com/dunstone-a/multistateQTL/issues}
}

}
\author{
\strong{Maintainer}: Amelia Dunstone \email{amelia.dunstone@svi.edu.au} (\href{https://orcid.org/0009-0009-6426-1529}{ORCID}) [contributor]

Authors:
\itemize{
  \item Christina Del Azodi \email{cazodi@svi.edu.au}
}

Other contributors:
\itemize{
  \item Davis McCarthy \email{dmccarthy@svi.edu.au} [contributor]
}

}
\keyword{internal}
