% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNAs.R
\name{replaceNAs}
\alias{replaceNAs}
\title{Return multistateQTL object with NAs filled in}
\usage{
replaceNAs(
  object,
  methods = list(betas = 0, errors = "mean", pvalues = 1, lfsrs = 1),
  verbose = FALSE
)
}
\arguments{
\item{object}{A `QTLExperiment` object}

\item{methods}{A named list with the method for each assay. Available methods
are to replace with a given integer or with the row mean or 
median.}

\item{verbose}{logical. Whether to print progress messages.}
}
\description{
A convenience function for imputing or filling in NAs in a `QTLExperiment` 
object.
}
